refactor to flat repo structure, test passing
