+++ atomate2_openmm/schemas/physical_state.py
