+++ atomate2_openmm/schemas/task_details.py
