+++ setup.py
@@ -1,9 +1,8 @@
-from setuptools import setup, find_namespace_packages
+from setuptools import setup, find_packages
 
 
 setup(
     name="atomate2-openmm",
-    packages=find_namespace_packages(where="src", include=["atomate2.*"]),
+    packages=find_packages(),
     version="0.0.2",
-    package_dir={"": "src"},
 )
