+++ tests/jobs/test_openmm_set_maker.py
@@ -1,0 +1,14 @@
+def test_openmm_set_maker(job_store):
+    from atomate2_openmm.jobs.openmm_set_maker import OpenMMSetMaker
+    from jobflow import run_locally
+
+    input_mol_dicts = [
+        {"smile": "O", "count": 200},
+        {"smile": "CCO", "count": 20},
+    ]
+
+    set_maker = OpenMMSetMaker()
+
+    set_job = set_maker.make(input_mol_dicts=input_mol_dicts, density=1)
+
+    responses = run_locally(set_job, store=job_store)
